feat(build): Add Git archives version files (#635)
When downloading the source code of a branch or a release from GitHub or
the mirror, there is no Git version information, making setuptools-scm
unable to read the version. The alternative way is to add a git archive
file supported by setuptools-scm. See:
* https://setuptools-scm.readthedocs.io/en/latest/usage/#git-archives
(available in setuptools-scm 7.0.0)
*
https://git-scm.com/docs/git-archive#Documentation/git-archive.txt-export-subst
*
https://docs.github.com/en/repositories/working-with-files/using-files/downloading-source-code-archives

Generated by the task: https://github.com/njzjz-bot/njzjz-bot/issues/4.